[relops] Divide algorithms into interface and implementation parts.
covered_by, intersects, touches, within.
